requirements.txt
SQLAlchemy~=2.0.41
Alp~=0.1.1
openai~=1.97.0
networkx~=3.5
streamlit~=1.47.0
python-dotenv~=1.1.1
fastapi~=0.116.1
pydantic~=2.11.7
structlog>=24.1.0
opentelemetry-sdk>=1.27.0
opentelemetry-api>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-logging>=0.48b0
pytest~=8.4.1
httpx==0.27.2